filter_state: rename files
This commit is renaming the file `socket_option_cilium_policy.h` to
`filter_state_cilium_policy.h`.
