+++ cilium/bpf_metadata.cc
@@ -41,12 +41,12 @@
 #include "cilium/api/bpf_metadata.pb.h"
 #include "cilium/api/bpf_metadata.pb.validate.h" // IWYU pragma: keep
 #include "cilium/conntrack.h"
+#include "cilium/filter_state_cilium_policy.h"
 #include "cilium/host_map.h"
 #include "cilium/ipcache.h"
 #include "cilium/network_policy.h"
 #include "cilium/policy_id.h"
 #include "cilium/socket_option_cilium_mark.h"
-#include "cilium/socket_option_cilium_policy.h"
 #include "cilium/socket_option_ip_transparent.h"
 
 namespace Envoy {
