+++ cilium/bpf_metadata.h
@@ -19,11 +19,11 @@
 #include "absl/types/optional.h"
 #include "cilium/api/bpf_metadata.pb.h"
 #include "cilium/conntrack.h"
+#include "cilium/filter_state_cilium_policy.h"
 #include "cilium/host_map.h"
 #include "cilium/ipcache.h"
 #include "cilium/network_policy.h"
 #include "cilium/socket_option_cilium_mark.h"
-#include "cilium/socket_option_cilium_policy.h"
 #include "cilium/socket_option_source_address.h"
 
 namespace Envoy {
