+++ cilium/filter_state_cilium_policy.cc
@@ -1,4 +1,4 @@
-#include "cilium/socket_option_cilium_policy.h"
+#include "cilium/filter_state_cilium_policy.h"
 
 #include <asm-generic/socket.h>
 #include <netinet/in.h>
